Add "https://" before movierus.tk
Without the "https://" before the website address, github would think that there is a file in the repo called 'movierus.tk' and when the user clicks on the hyperlink, github would give the user a 404 error. Hopefully, by identifying it as a website to github, it should now redirect to the website and not an 404 not found page from github.