Remove tparser.org
The site seems to be down for a long time now (more than a year).
It had a mirror **tparser.me** which is also down for a long time now. Then there was kind of a clone **Riptide.cz** that is also down.
All those sites either not available or redirect to some scamming/phishing sites.